Refatora montagem de HTML provisório da copy em resolver e processor
